Uploaded a python script with documentation and problem statement
Q.Find the best possible node to poison the graph with fake news assuming that a person reveals the information to a person he/she likes with some probability p, where 0<p<1
Soln. We have used a probabilistic approach for finding the best possible node using a modified BFS algorithm.