Made our program more generic
We have now considered that each person likes another person with a certain degree, so he or she will tell the news with a different probability, which we have randomly generated, so our program can now handle such a problem too. Also, earlier we were just printing the best node to spread a news, but, now we are printing the whole ranking of nodes according to their news spreading capability best to worst.